pyproject.toml
[project]
name = "flask-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-pymongo>=3.0.1",
    "motor>=3.7.0",
    "passlib>=1.7.4",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
